migrate to spirv.hpp11
make spirv-cross compiled using std=c++17 (fix issue https://github.com/KhronosGroup/SPIRV-Cross/issues/2523)
Need more support for fix this in other platform

+++ spirv_cross_containers.hpp
@@ -642,7 +642,7 @@ public:
 	template <typename T, typename std::enable_if<!std::is_floating_point<T>::value, int>::type = 0>
 	StringStream &operator<<(const T &t)
 	{
-		auto s = std::to_string(t);
+		auto s = std::to_string(static_cast<uint32_t>(t));
 		append(s.data(), s.size());
 		return *this;
 	}
